+++ src/Composer/Installers/QuicksilverInstaller.php
@@ -1,0 +1,9 @@
+<?php
+namespace Composer\Installers;
+
+class QuicksilverInstaller extends BaseInstaller
+{
+    protected $locations = array(
+        'script'    => 'web/private/scripts/quicksilver/{$name}/',
+    );
+}